+++ iBXTest/iBXTest/Sources/Controllers/NavigationTestController/NavigationPanelTestController.m
@@ -44,6 +44,14 @@
     return result;
 }
 
+- (UIView*) navigationBackgroundWithController: (BxNavigationController*) navigationController
+{
+    UIImageView * imageView = [[UIImageView alloc] initWithImage: [UIImage imageNamed: @"common_background1.jpg"]];
+    imageView.contentMode = UIViewContentModeScaleAspectFill;
+    imageView.clipsToBounds = YES;
+    return imageView;
+}
+
 - (void) viewWillLayoutSubviews
 {
     [super viewWillLayoutSubviews];